Remove ANTLR3 parser.
